Fixes required to begin porting tests over
`runtime::test` requires the function be `Send`, and we also couldn't
chain the futures from `h2-support` since `Box<dyn Future>` requires
`Unpin` (or the alternative that we've taken, which is pinning the box).

Additionally, my reading of `Mock`'s impls being the same as `Pipe`s was
wrong, as they read and write to different buffers and use different
wakers.