Port `tests/server.rs` to `std::future`
The tests were mostly mechanical to port, but I've attempted to do some
additional cleanup. A lot of tests were ending with
`srv.into_future().map(|_| ())`, as a proxy for "continue polling until
the stream is closed". These tests probably should have at least been
asserting they got no more requests, but I've taken it a step further
and actually provided a future for exactly that purpose.

From my reading of the code, my understanding is that calling
`poll_close` in a loop without ever checking to see if we got another
request is implicitly asserting there are no more requests, since it
should never close when there's an incoming request pending, or at the
very least it should only happen if the client sent a request and then
closed the connection without expecting any response back.

I've added a similar future to `SendStream`, and intend to continue
doing this any time we're using `poll_fn` in our tests, as having a
`poll_x` function on anything is a pretty strong signal that we want a
version of it that returns a future as well (I suspect the reason that
this was avoided was that that taking `&mut self` and having the
lifetime on the future can cause ownerhsip issues in futures 0.1, but
async/await makes this much easier)

I've opted to use the runtime crate here, as it seemed like the lowest
friction way to get `async fn` for tests. It's a bit more heavyweight
than we need, and it has the downside of requiring `Send` (though that
was probably something we should have done anyway). Ultimately all we
really need is something that wraps the whole test in `block_on(async {
... })`, but the flexibility may be nice later if we ever switch to
actually using tokio's executor here.